requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.3.1
anyio==4.2.0
astroid==2.5.1
async-timeout==3.0.1
attrs==21.4.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.3.2
colorama==0.4.4
cryptography==3.4.7
discord==2.3.2
discord.py==2.3.2
exceptiongroup==1.2.0
ffmpeg==1.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.3
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==6.0.2
mutagen==1.47.0
pillow==10.2.0
pycparser==2.20
pycryptodomex==3.19.1
pylint==2.7.2
pymongo==3.12.1
PyNaCl==1.5.0
pyspnego==0.1.5
requests==2.31.0
six==1.15.0
smbprotocol==1.5.0
sniffio==1.3.0
toml==0.10.2
typing_extensions==4.1.1
urllib3==2.1.0
websockets==12.0
wrapt==1.12.1
yarl==1.7.2
youtube-dl==2021.12.17
youtube-search-python==1.6.6
yt-dlp==2023.12.30
